+++ R/07makeBibLatexAuthoryear.R
@@ -1,39 +1,7 @@
 #' @keywords internal
-MakeAuthorYear <- function(docstyle = "text") local({
-  docstyle <- get("docstyle", parent.frame(2))
+MakeAuthorYear <- function(docstyle = "text"){
 ##################################################################
 ## Formatting functions
-
-cleanupLatex <- function (x){
-  if (!length(x)) 
-    return(x)
-  
-  if (any(grepl('mkbib', x, useBytes = TRUE))){
-    x <- gsub('mkbibquote', 'dQuote', x, useBytes = TRUE)
-    x <- gsub('mkbibemph', 'emph', x, useBytes = TRUE)
-    x <- gsub('mkbibbold', 'bold', x, useBytes = TRUE)
-  }
-  x <- gsub('\\\\hyphen', '-', x, useBytes = TRUE)
-  
-  latex <- try(tools::parseLatex(x), silent = TRUE)
-  if (inherits(latex, "try-error")) {
-    x
-  }else{
-    x <- tools::deparseLatex(tools::latexToUtf8(latex), dropBraces = TRUE)
-    if (grepl("\\\\[[:punct:]]", x, useBytes = TRUE)){
-      x <- gsub("\\\\'I", '\u00cd', x, useBytes = TRUE)
-      x <- gsub("\\\\'i", '\u00ed', x, useBytes = TRUE)
-      x <- gsub('\\\\"I', '\u00cf', x, useBytes = TRUE)
-      x <- gsub('\\\\"i', '\u00ef', x, useBytes = TRUE)
-      x <- gsub("\\\\\\^I", '\u00ce', x, useBytes = TRUE)
-      x <- gsub("\\\\\\^i", '\u00ee', x, useBytes = TRUE)
-      x <- gsub("\\\\`I", '\u00cc', x, useBytes = TRUE)
-      x <- gsub("\\\\`i", '\u00ec', x, useBytes = TRUE)
-      Encoding(x) <- 'UTF-8'
-    }
-    x
-  }
-}
 
 collapse <- function(strings){ 
   paste(strings, collapse = "\n")
@@ -670,11 +638,12 @@ fmtEditor <- function(doc, editor.used.already = FALSE, prefix = NULL, suffix = 
 }
 
 fmtJTitle <- function(title){
-  if (grepl('[.?!]$', title, useBytes = TRUE)){
-    paste0("\\dQuote{", collapse(cleanupLatex(title)), "}")
-  }else{
-    paste0("\\dQuote{", collapse(cleanupLatex(title)), "}.")
-  }
+  if (!is.null(title))  
+    if (grepl('[.?!]$', title, useBytes = TRUE)){
+      paste0("\\dQuote{", collapse(cleanupLatex(title)), "}")
+    }else{
+      paste0("\\dQuote{", collapse(cleanupLatex(title)), "}.")
+    }
 }
 
 fmtBTitle <- function(tl, stl){
@@ -1144,4 +1113,4 @@ formatUnpublished <- function(paper){
 }
 
 environment()
-})
+}